🚚 rename: index파일, style파일 이름 변경
