+++ services/quiz/src/components/StudyDetailPage/QuizList/index.tsx
