+++ services/quiz/src/components/StudyDetailPage/Ranking/index.tsx
