+++ services/quiz/src/components/StudyDetailPage/StudyInfo/index.tsx
