+++ services/quiz/src/components/StudyDetailPage/StudyProgress/index.tsx
