+++ services/quiz/src/components/common/Header/ProfileIcon/index.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 
 import ic_default_profile from '../../../../img/ic_default_profile.png';
-import { ProfileIconCss } from './style/ProfileIconCss';
+import { ProfileIconCss } from './style';
 
 function ProfileIcon() {
   return <Image css={ProfileIconCss} src={ic_default_profile} alt="사용자 프로필 이미지" />;
