+++ services/quiz/src/components/common/Header/ProfileIcon/style.ts
